adds a failing test for the getLink function in the cli_test.go file and changes the function in the cli.go file to return an error if the clipboard operation errors out

+++ cli_test.go
@@ -1,7 +1,6 @@
 package main
 
 import (
-	"fmt"
 	"os"
 	"path"
 	"testing"
@@ -10,12 +9,22 @@ import (
 var testDir = path.Join("./fixtures", "Links.json")
 
 func TestStartCli(t *testing.T) {
-	// test add Link
 
+	// test add Link
 	_, err := StartCli([]string{"./scribe", "al", "goog", "www.google.com"}, testDir)
 	if err != nil {
-		fmt.Printf("The addLink command encountered the follwing error: %s", err)
-		t.Fail()
+		t.Errorf("The addLink command encountered the following error: %s", err)
+	}
+
+	// test get Link
+	expected := "www.google.com"
+	resp, err := StartCli([]string{"./scribe", "gl", "goog"}, testDir)
+	if err != nil {
+		t.Errorf("The getLink command encountered the following error: %s", err)
+	}
+
+	if resp[0] != expected {
+		t.Error("The getLink command did not return the expected output")
 	}
 
 	err = os.RemoveAll(testDir)
